Saving work: separated pieces but still needs work with redux.
