+++ src/util/style.tsx
@@ -1,3 +1,3 @@
 export default {
     
-}
+};
